# src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: SpongeblogSP API
  description: Spongeblog blogging API
  version: "1.0.0"
# array of all schemes that your API supportsf
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
parameters:
...
  BlogEdit:
    name: blog-edit
    in: body
    schema:
      $ref: '#/definitions/BlogDTO'
    required: true
...
paths:
...
  /api/v1/blog:
    post:
      summary: Create Blog
      description: Creates new Blog entry.
      parameters:
        - $ref: '#/parameters/BlogEdit'
      tags: 
        - Blog
      responses:
        201:
          description: Created Blog.
          schema:
            $ref: '#/definitions/BlogDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    put:
      summary: Update Blog
      description: Updated a Blog entry.
      parameters:
        - $ref: '#/parameters/BlogEdit'
      tags: 
        - Blog
      responses:
        200:
          description: Blog updated.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    get:
      summary: Get Blog
      description: Returns the Blog entry.
      tags: 
        - Blog
      responses:
        200:
          description: Blog.
          schema:
            $ref: '#/definitions/BlogDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    delete:
      summary: Delete Blog
      description: Deletes existing Blog.
      tags: 
        - Blog
      responses:
        200:
          description: Blog deleted.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/blog/rewind:
    put:
      summary: Rewind Blog
      description: Restores previous Blog entry.
      tags: 
        - Blog
      responses:
        200:
          description: Previous Blog restored.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
...
definitions:
...
  BlogDTO:
    type: object
    properties:
      name:
        type: string
        description: Name of the Blog.
      status:
        type: string
        description: Status of the Blog. 
        enum: 
          - ACTIVE
          - DISABLED
          - MAINTENANCE
      settings:
        type: object
        description: Settings map of the Blog.
        additionalProperties:
          type: string
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string